feat(sln): .net 8, xperience 28.1, CPM, formatting
