+++ global.json
@@ -1,7 +1,7 @@
 {
   "sdk": {
-    "version": "7.0.203",
-    "rollForward": "latestMinor",
+    "version": "8.0.100",
+    "rollForward": "latestMajor",
     "allowPrerelease": false
   }
 }
