--- test/XperienceCommunity.PreviewComponentOutlinesTests/XperienceCommunity.PreviewComponentOutlinesTests.csproj
@@ -1,29 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <IsPackable>false</IsPackable>
-    <IsTestProject>true</IsTestProject>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.5.0" />
-    <PackageReference Include="NUnit" Version="3.13.3" />
-    <PackageReference Include="NUnit3TestAdapter" Version="4.4.2" />
-    <PackageReference Include="NUnit.Analyzers" Version="3.6.1" />
-    <PackageReference Include="coverlet.collector" Version="3.2.0" />
-    <PackageReference Include="Kentico.Xperience.Core.Tests" Version="25.0.0" />
-    <PackageReference Include="FluentAssertions" Version="6.11.0" />
-    <PackageReference Include="NSubstitute" Version="5.0.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Using Include="NUnit.Framework" />
-    <Using Include="NSubstitute" />
-    <Using Include="XperienceCommunity.PreviewComponentOutlinesTests" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\XperienceCommunity.PreviewComponentOutlines\XperienceCommunity.PreviewComponentOutlines.csproj" />
-  </ItemGroup>
-
-</Project>
