+++ tests/XperienceCommunity.PreviewComponentOutlines.Tests/XperienceCommunity.PreviewComponentOutlines.Tests.csproj
@@ -1,0 +1,29 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <IsPackable>false</IsPackable>
+    <IsTestProject>true</IsTestProject>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" />
+    <PackageReference Include="NUnit" />
+    <PackageReference Include="NUnit3TestAdapter"/>
+    <PackageReference Include="NUnit.Analyzers"/>
+    <PackageReference Include="coverlet.collector"/>
+    <PackageReference Include="Kentico.Xperience.Core.Tests" />
+    <PackageReference Include="FluentAssertions" />
+    <PackageReference Include="NSubstitute"/>
+  </ItemGroup>
+
+  <ItemGroup>
+    <Using Include="NUnit.Framework" />
+    <Using Include="NSubstitute" />
+    <Using Include="XperienceCommunity.PreviewComponentOutlines.Tests" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\XperienceCommunity.PreviewComponentOutlines\XperienceCommunity.PreviewComponentOutlines.csproj" />
+  </ItemGroup>
+
+</Project>
